requirements.txt
attrs==21.4.0
click==8.1.1
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.1.1
python-dotenv==0.20.0
SQLAlchemy==1.4.32
tomli==2.0.1
Werkzeug==2.1.0
zipp==3.7.0
